+++ integrations/github/pr.go
@@ -1,10 +1,9 @@
-// Structs and methods used to process a pull request.
+// Package github integrates everything necessary to test commits, comment on
+// pull requests and close them if the build failed.
 package github
 
 import (
-	"encoding/base64"
 	"encoding/json"
-	"io/ioutil"
 	"leeroy/config"
 	"leeroy/logging"
 	"log"
@@ -14,31 +13,35 @@ import (
 
 // TODO: parse full body, not just the fields needed
 
+// PRCallback handles pull requests coming from GitHubs webhook.
 type PRCallback struct {
 	Number int
 	Action string
 	PR     PRPullRequest `json:"pull_request"`
 }
 
+// PRPullRequest stores the most basic information about a pull request.
 type PRPullRequest struct {
-	Url          string
-	State        string
-	Comments_url string
-	Head         PRCommit
+	URL         string `json:"url"`
+	State       string
+	CommentsURL string `json:"comments_url"`
+	Head        PRCommit
 }
 
+// PRCommit points to the latest commit and repository of a pull request.
 type PRCommit struct {
 	Commit     string `json:"sha"`
 	Repository PRRepo `json:"repo"`
 }
 
+// PRRepo stores the repository URL of a pull request.
 type PRRepo struct {
-	Html_url string `json:"html_url"`
+	HTMLURL string `json:"html_url"`
 }
 
-// Returns base URL for repository (HTML, not API)
+// RepoURL returns the base URL for repository (HTML, not API)
 func (p *PRCallback) RepoURL() string {
-	return p.PR.Head.Repository.Html_url
+	return p.PR.Head.Repository.HTMLURL
 }
 
 // Handle GitHub pull requests.
@@ -49,8 +52,7 @@ func handlePR(req *http.Request, blog *logging.Buildlog, c *config.Config) {
 	err := json.Unmarshal(b, &pc)
 
 	if err != nil {
-		log.Println(string(b))
-		panic("Could not unmarshal request")
+		log.Fatalln("Could not unmarshal PR request")
 	}
 
 	if pc.Action != "closed" {
@@ -70,8 +72,7 @@ func updatePR(pc PRCallback, blog *logging.Buildlog, c *config.Config) {
 				r, err := c.ConfigForRepo(j.URL)
 
 				if err != nil {
-					log.Println(err)
-					return
+					log.Fatalln(err)
 				}
 
 				if r.CommentPR {
@@ -94,7 +95,7 @@ func updatePR(pc PRCallback, blog *logging.Buildlog, c *config.Config) {
 			}
 			counter = 0
 		} else {
-			counter += 1
+			counter++
 		}
 
 		time.Sleep(10 * time.Second)
@@ -103,42 +104,19 @@ func updatePR(pc PRCallback, blog *logging.Buildlog, c *config.Config) {
 
 // Returns if PRCallback is for the latest commit.
 func prIsCurrent(pc PRCallback, c *config.Config) bool {
-	cl := &http.Client{}
-	r, err := http.NewRequest("GET", pc.PR.Url, nil)
-
-	if err != nil {
-		log.Println(err)
-		return true
-	}
-
 	rp, err := c.ConfigForRepo(pc.RepoURL())
 
-	if err != nil {
-		log.Println(err)
-		return true
-	}
-
-	t := base64.URLEncoding.EncodeToString([]byte(rp.AccessKey))
-
-	r.Header.Add("content-type", "application/json")
-	r.Header.Add("Authorization", "Basic "+t)
-
-	resp, err := cl.Do(r)
+	r, err := githubRequest("GET", pc.PR.URL, rp.AccessKey, nil)
 
 	if err != nil {
-		log.Println(err)
-		return true
+		log.Fatalln(err)
 	}
-
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
 
 	var pr PRPullRequest
-	err = json.Unmarshal(body, &pr)
+	err = json.Unmarshal(r, &pr)
 
 	if err != nil {
-		log.Println(err)
-		return true
+		log.Fatalln(err)
 	}
 
 	if pr.Head.Commit != pc.PR.Head.Commit {