+++ notification/notification.go
@@ -1,39 +1,66 @@
-// Notifications to send after a build process finished.
+// Package notification handles all notifications for a job. This includes
+// build and deployment notifications.
 package notification
 
 import (
+	"bytes"
 	"leeroy/config"
 	"leeroy/logging"
 	"log"
+	"text/template"
 )
 
-// Run all notifications configured for a
-func Notify(c *config.Config, j *logging.Job) {
-	// always notify the person who comitted
-	go email(c, j, j.Email)
+// Define default template names that are used for various notifications.
+const (
+	KindBuild       = "build"
+	KindDeployStart = "deploy-start"
+	KindDeployDone  = "deploy-done"
+)
 
-	repo, err := c.ConfigForRepo(j.URL)
+// List of all valid notification types.
+var kinds = [...]string{KindBuild, KindDeployStart, KindDeployDone}
+
+// Everything related to a notification.
+type notification struct {
+	Repo    string
+	Branch  string
+	Name    string
+	Email   string
+	Status  bool
+	Deploy  bool
+	URL     string
+	kind    string
+	message string
+}
+
+// Create a notification from a job.
+func notificationFromJob(j *logging.Job, c *config.Config) *notification {
+	return &notification{
+		Repo:   j.URL,
+		Branch: j.Branch,
+		Name:   j.Name,
+		Email:  j.Email,
+		Status: j.Success(),
+		URL:    j.StatusURL(c.URL),
+		Deploy: j.DeploySuccess(),
+	}
+}
+
+// Render a notification.
+func (n *notification) render() {
+	t := template.New(n.kind)
+	t, err := t.Parse(templates[n.kind])
 
 	if err != nil {
-		log.Println("could not find repo", j.URL)
-		return
+		log.Fatal(err)
 	}
 
-	for _, n := range repo.Notify {
-		switch n.Service {
-		case "email":
-			// Arguments for email are the mail addresses to notify
-			for mail, _ := range n.Arguments {
-				go email(c, j, mail)
-			}
-		case "slack":
-			go slack(c, j, n.Arguments["endpoint"], n.Arguments["channel"])
-		case "hipchat":
-			go hipchat(c, j, n.Arguments["key"], n.Arguments["channel"])
-		case "campfire":
-			go campfire(c, j, n.Arguments["id"], n.Arguments["room"], n.Arguments["key"])
-		default:
-			log.Println("Notification not supported", n.Service)
-		}
+	var r bytes.Buffer
+	err = t.Execute(&r, n)
+
+	if err != nil {
+		log.Fatal(err)
 	}
+
+	n.message = r.String()
 }